+++ extensions/brand-dev/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Brand.dev Changelog
 
-## [Proper Case Address Items] - {PR_MERGE_DATE}
+## [Open Multiple Logos] - {PR_MERGE_DATE}
 
 - In `Detail` view the proper Address item title is shown e.g. "Country Code" instead of "country_code"
+- You can now open more than the first logo in browser e.g. if a Brand has multiple logos, you can open the 2nd logo instead of first
 
 ## [Choose Preferred Secondary Action] - 2024-11-24